refactor: testFixtures 폴더로 textFixture 이관
