+++ src/test/java/persistence/testFixtures/PkHasPerson.java
@@ -1,11 +1,16 @@
-package persistence.sql.ddl;
+package persistence.testFixtures;
 
+import jakarta.persistence.Entity;
 import jakarta.persistence.Id;
 
-public class NoHasEntity {
+@Entity
+public class PkHasPerson {
 
     @Id
     private Long id;
+
     private String name;
+
     private Integer age;
+
 }
